+++ src/components/app/app.scss
@@ -1,4 +1,5 @@
 @use './theme.scss';
+@use '../ui-kit/mixins.scss';
 
 html,
 body {
@@ -9,6 +10,7 @@ body {
 body {
   background-color: var(--background-base);
   color: var(--text-base);
+  font-family: sans-serif;
 }
 
 @for $i from 1 through 10 {
@@ -35,4 +37,8 @@ body {
 
 a {
   color: var(--link);
+}
+
+.link-button {
+  @include mixins.button;
 }
