+++ src/components/ui-kit/mixins.scss
@@ -1,0 +1,20 @@
+@mixin button() {
+  display: inline-flex;
+  border: none;
+  padding: 8px 16px;
+  border-radius: 4px;
+  background-color: var(--button-background);
+  color: var(--text-primary);
+  text-decoration: none;
+  font-size: 1rem;
+
+  &:hover {
+    cursor: pointer;
+    background-color: var(--button-background-active);
+  }
+
+  &.active {
+    cursor: default;
+    background-color: var(--button-background-active);
+  }
+}
